agent-evaluation/requirements.txt
# AI Observability - TruLens Framework
trulens-core>=2.1.2
trulens-connectors-snowflake>=2.1.2
trulens-providers-cortex>=2.1.2

# Snowflake Python
snowflake-snowpark-python>=1.11.0

# HTTP Requests (for Cortex Agent API)
requests>=2.31.0

# Data Processing
pandas>=2.0.0
